+++ src/com/nixinova/input/Coord.java
@@ -1,0 +1,19 @@
+package com.nixinova.input;
+
+public class Coord {
+	
+	public int x;
+	public int y;
+	public int z;
+	
+	public Coord() {
+		this.x = this.y = this.z = 0;
+	}
+	
+	public Coord(int x, int y, int z) {
+		this.x = x;
+		this.y = y;
+		this.z = z;
+	}
+
+}