+++ src/com/nixinova/main/Display.java
@@ -127,15 +127,9 @@ public class Display extends Canvas implements Runnable {
 		graphics.drawImage(this.img, 0, 0, WIDTH, HEIGHT, null);
 		graphics.setColor(Color.white);
 
-
 		String playerX = String.format("%01d", Math.round(Controller.playerX));
 		String playerY = String.format("%01d", Math.round(Controller.playerY));
 		String playerZ = String.format("%01d", Math.round(Controller.playerZ));
-
-		String intLimit = String.valueOf(Integer.MAX_VALUE);
-		playerX = playerX.contains(intLimit) ? "infinity" : playerX;
-		playerY = playerY.contains(intLimit) ? "infinity" : playerY;
-		playerZ = playerZ.contains(intLimit) ? "infinity" : playerZ;
 
 		String msg1 = "", msg2 = "", msg3 = "";
 		final int OPTIONS_VERSION = Options.OPTIONS_VERSION;
@@ -152,7 +146,7 @@ public class Display extends Canvas implements Runnable {
 		}
 
 		if (Controller.debugShown) {
-			int sep = 15;
+			final int sep = 15;
 			graphics.setColor(Color.white);
 			graphics.drawString(Mineo.TITLE, 5, sep);
 			graphics.drawString("FPS: " + String.valueOf(this.fps), 5, sep * 2);